MAINT: refactored code for fmri connectome interface
