+++ cmtklib/connectome.py
@@ -254,333 +254,6 @@ def probtrackx_cmat(voxel_connectivity_files, roi_volumes, parcellation_scheme, 
                 g2.node[u_gml]['dn_region'] = d_gml['dn_region']
             nx.write_graphml(g2,'connectome_%s.graphml' % parkey)
 
-def prob_cmat(intrk, roi_volumes, parcellation_scheme, output_types=['gPickle'], atlas_info = {}):
-
-    print("Filling probabilistic connectivity matrices:")
-
-    if parcellation_scheme != "Custom":
-        resolutions = get_parcellation(parcellation_scheme)
-    else:
-        resolutions = atlas_info
-
-    firstROIFile = roi_volumes[0]
-    firstROI = nibabel.load(firstROIFile)
-    roiVoxelSize = firstROI.get_header().get_zooms()
-
-    for parkey, parval in resolutions.items():
-        print("Resolution = "+parkey)
-
-        # Open the corresponding ROI
-        print("Open the corresponding ROI")
-        for vol in roi_volumes:
-            if parkey in vol:
-                roi_fname = vol
-                print roi_fname
-        roi       = nibabel.load(roi_fname)
-        roiData   = roi.get_data()
-
-        # Create the matrix
-        nROIs = parval['number_of_regions']
-        print("Create the connection matrix (%s rois)" % nROIs)
-        G     = nx.Graph()
-
-        # add node information from parcellation
-        gp = nx.read_graphml(parval['node_information_graphml'])
-        for u,d in gp.nodes(data=True):
-            G.add_node(int(u))
-            for key in d:
-                G.node[int(u)][key] = d[key]
-            # compute a position for the node based on the mean position of the
-            # ROI in voxel coordinates (segmentation volume )
-            G.node[int(u)]['dn_position'] = tuple(np.mean( np.where(roiData== int(d["dn_correspondence_id"]) ) , axis = 1))
-
-        graph_matrix = np.zeros((nROIs,nROIs),dtype = int)
-
-        pc = -1
-
-        for intrk_i in range(0,len(intrk)):
-            pcN = int(round( float(100*intrk_i)/len(intrk) ))
-            if pcN > pc and pcN%20 == 0:
-                pc = pcN
-                print('%4.0f%%' % (pc))
-            fib, hdr    = nibabel.trackvis.read(intrk[intrk_i], False)
-            (endpoints,endpointsmm) = create_endpoints_array(fib, roiVoxelSize, False)
-            n = len(fib)
-
-            dis = 0
-
-            pc = -1
-            for i in range(n):  # n: number of fibers
-
-                # ROI start => ROI end
-                try:
-                    startROI = int(roiData[endpoints[i, 0, 0], endpoints[i, 0, 1], endpoints[i, 0, 2]]) # endpoints from create_endpoints_array
-                    endROI   = int(roiData[endpoints[i, 1, 0], endpoints[i, 1, 1], endpoints[i, 1, 2]])
-                except IndexError:
-                    print("An index error occured for fiber %s. This means that the fiber start or endpoint is outside the volume. Continue." % i)
-                    continue
-
-                # Filter
-                if startROI == 0 or endROI == 0:
-                    dis += 1
-                    #fiberlabels[i,0] = -1
-                    continue
-
-                if startROI > nROIs or endROI > nROIs:
-        #                print("Start or endpoint of fiber terminate in a voxel which is labeled higher")
-        #                print("than is expected by the parcellation node information.")
-        #                print("Start ROI: %i, End ROI: %i" % (startROI, endROI))
-        #                print("This needs bugfixing!")
-                    continue
-
-                # Update fiber label
-                # switch the rois in order to enforce startROI < endROI
-                if endROI < startROI:
-                    tmp = startROI
-                    startROI = endROI
-                    endROI = tmp
-
-                # Add edge to graph
-                if G.has_edge(startROI, endROI):
-                    G[startROI][endROI]['n_tracks'] += 1
-                else:
-                    G.add_edge(startROI, endROI, n_tracks  = 1)
-
-                graph_matrix[startROI-1][endROI-1] +=1
-
-        tot_tracks = graph_matrix.sum()
-
-        for u,v,d in G.edges(data=True):
-            G.remove_edge(u,v)
-            di = { 'number_of_fibers' : (float(d['n_tracks']) / tot_tracks.astype(float))}
-            G.add_edge(u,v)
-            for key in di:
-                G[u][v][key] = di[key]
-
-        # storing network
-        if 'gPickle' in output_types:
-            nx.write_gpickle(G, 'connectome_%s.gpickle' % parkey)
-        if 'mat' in output_types:
-            # edges
-            size_edges = (parval['number_of_regions'],parval['number_of_regions'])
-            edge_keys = G.edges(data=True)[0][2].keys()
-
-            edge_struct = {}
-            for edge_key in edge_keys:
-                edge_struct[edge_key] = nx.to_numpy_matrix(G,weight=edge_key)
-
-            # nodes
-            size_nodes = parval['number_of_regions']
-            node_keys = G.nodes(data=True)[0][1].keys()
-
-            node_struct = {}
-            for node_key in node_keys:
-                if node_key == 'dn_position':
-                    node_arr = np.zeros([size_nodes,3],dtype=np.float)
-                else:
-                    node_arr = np.zeros(size_nodes,dtype=np.object_)
-                node_n = 0
-                for _,node_data in G.nodes(data=True):
-                    node_arr[node_n] = node_data[node_key]
-                    node_n += 1
-                node_struct[node_key] = node_arr
-
-            scipy.io.savemat('connectome_%s.mat' % parkey, mdict={'sc':edge_struct,'nodes':node_struct})
-        if 'graphml' in output_types:
-            g2 = nx.Graph()
-            for u_gml,v_gml,d_gml in G.edges(data=True):
-                g2.add_edge(u_gml,v_gml)
-                for key in d_gml:
-                    g2[u_gml][v_gml][key] = d_gml[key]
-            for u_gml,d_gml in G.nodes(data=True):
-                g2.add_node(u_gml)
-                g2.node[u_gml]['dn_correspondence_id'] = d_gml['dn_correspondence_id']
-                g2.node[u_gml]['dn_fsname'] = d_gml['dn_fsname']
-                g2.node[u_gml]['dn_hemisphere'] = d_gml['dn_hemisphere']
-                g2.node[u_gml]['dn_name'] = d_gml['dn_name']
-                g2.node[u_gml]['dn_position_x'] = d_gml['dn_position'][0]
-                g2.node[u_gml]['dn_position_y'] = d_gml['dn_position'][1]
-                g2.node[u_gml]['dn_position_z'] = d_gml['dn_position'][2]
-                g2.node[u_gml]['dn_region'] = d_gml['dn_region']
-            nx.write_graphml(g2,'connectome_%s.graphml' % parkey)
-
-# class MRTrix3BaseInputSpec(CommandLineInputSpec):
-#     nthreads = traits.Int(
-#         argstr='-nthreads %d', desc='number of threads. if zero, the number'
-#         ' of available cpus will be used', nohash=True)
-#     # DW gradient table import options
-#     grad_file = File(exists=True, argstr='-grad %s',
-#                      desc='dw gradient scheme (MRTrix format')
-#     grad_fsl = traits.Tuple(
-#         File(exists=True), File(exists=True), argstr='-fslgrad %s %s',
-#         desc='(bvecs, bvals) dw gradient scheme (FSL format')
-#     bval_scale = traits.Enum(
-#         'yes', 'no', argstr='-bvalue_scaling %s',
-#         desc='specifies whether the b - values should be scaled by the square'
-#         ' of the corresponding DW gradient norm, as often required for '
-#         'multishell or DSI DW acquisition schemes. The default action '
-#         'can also be set in the MRtrix config file, under the '
-#         'BValueScaling entry. Valid choices are yes / no, true / '
-#         'false, 0 / 1 (default: true).')
-
-#     in_bvec = File(exists=True, argstr='-fslgrad %s %s',
-#                    desc='bvecs file in FSL format')
-#     in_bval = File(exists=True, desc='bvals file in FSL format')
-
-
-# class MRTrix3Base(CommandLine):
-
-#     def _format_arg(self, name, trait_spec, value):
-#         if name == 'nthreads' and value == 0:
-#             value = 1
-#             try:
-#                 from multiprocessing import cpu_count
-#                 value = cpu_count()
-#             except:
-#                 logger.warn('Number of threads could not be computed')
-#                 pass
-#             return trait_spec.argstr % value
-
-#         if name == 'in_bvec':
-#             return trait_spec.argstr % (value, self.inputs.in_bval)
-
-#         return super(MRTrix3Base, self)._format_arg(name, trait_spec, value)
-
-#     def _parse_inputs(self, skip=None):
-#         if skip is None:
-#             skip = []
-
-#         try:
-#             if (isdefined(self.inputs.grad_file) or
-#                     isdefined(self.inputs.grad_fsl)):
-#                 skip += ['in_bvec', 'in_bval']
-
-#             is_bvec = isdefined(self.inputs.in_bvec)
-#             is_bval = isdefined(self.inputs.in_bval)
-#             if is_bvec or is_bval:
-#                 if not is_bvec or not is_bval:
-#                     raise RuntimeError('If using bvecs and bvals inputs, both'
-#                                        'should be defined')
-#                 skip += ['in_bval']
-#         except AttributeError:
-#             pass
-
-#         return super(MRTrix3Base, self)._parse_inputs(skip=skip)
-
-# class BuildConnectomeInputSpec(CommandLineInputSpec):
-#     in_file = File(exists=True, argstr='%s', mandatory=True, position=-3,
-#                    desc='input tractography')
-#     in_parc = File(exists=True, argstr='%s', position=-2,
-#                    desc='parcellation file')
-#     out_file = File(
-#         'connectome.csv', argstr='%s', mandatory=True, position=-1,
-#         usedefault=True, desc='output file after processing')
-
-#     nthreads = traits.Int(
-#         argstr='-nthreads %d', desc='number of threads. if zero, the number'
-#         ' of available cpus will be used', nohash=True)
-
-#     vox_lookup = traits.Bool(
-#         argstr='-assignment_voxel_lookup',
-#         desc='use a simple voxel lookup value at each streamline endpoint')
-#     search_radius = traits.Float(
-#         argstr='-assignment_radial_search %f',
-#         desc='perform a radial search from each streamline endpoint to locate '
-#         'the nearest node. Argument is the maximum radius in mm; if no node is'
-#         ' found within this radius, the streamline endpoint is not assigned to'
-#         ' any node.')
-#     search_reverse = traits.Float(
-#         argstr='-assignment_reverse_search %f',
-#         desc='traverse from each streamline endpoint inwards along the '
-#         'streamline, in search of the last node traversed by the streamline. '
-#         'Argument is the maximum traversal length in mm (set to 0 to allow '
-#         'search to continue to the streamline midpoint).')
-#     search_forward = traits.Float(
-#         argstr='-assignment_forward_search %f',
-#         desc='project the streamline forwards from the endpoint in search of a'
-#         'parcellation node voxel. Argument is the maximum traversal length in '
-#         'mm.')
-
-#     metric = traits.Enum(
-#         'count', 'meanlength', 'invlength', 'invnodevolume', 'mean_scalar',
-#         'invlength_invnodevolume', argstr='-metric %s', desc='specify the edge'
-#         ' weight metric')
-
-#     in_scalar = File(
-#         exists=True, argstr='-image %s', desc='provide the associated image '
-#         'for the mean_scalar metric')
-
-#     in_weights = File(
-#         exists=True, argstr='-tck_weights_in %s', desc='specify a text scalar '
-#         'file containing the streamline weights')
-
-#     keep_unassigned = traits.Bool(
-#         argstr='-keep_unassigned', desc='By default, the program discards the'
-#         ' information regarding those streamlines that are not successfully '
-#         'assigned to a node pair. Set this option to keep these values (will '
-#         'be the first row/column in the output matrix)')
-#     zero_diagonal = traits.Bool(
-#         argstr='-zero_diagonal', desc='set all diagonal entries in the matrix '
-#         'to zero (these represent streamlines that connect to the same node at'
-#         ' both ends)')
-
-
-# class BuildConnectomeOutputSpec(TraitedSpec):
-#     out_file = File(exists=True, desc='the output response file')
-
-
-# class BuildConnectome(MRTrix3Base):
-
-#     """
-#     Generate a connectome matrix from a streamlines file and a node
-#     parcellation image
-#     Example
-#     -------
-#     >>> import nipype.interfaces.mrtrix3 as mrt
-#     >>> mat = mrt.BuildConnectome()
-#     >>> mat.inputs.in_file = 'tracks.tck'
-#     >>> mat.inputs.in_parc = 'aparc+aseg.nii'
-#     >>> mat.cmdline                               # doctest: +ELLIPSIS +ALLOW_UNICODE
-#     'tck2connectome tracks.tck aparc+aseg.nii connectome.csv'
-#     >>> mat.run()                                 # doctest: +SKIP
-#     """
-
-#     _cmd = 'tck2connectome'
-#     input_spec = BuildConnectomeInputSpec
-#     output_spec = BuildConnectomeOutputSpec
-
-#     def _list_outputs(self):
-#         outputs = self.output_spec().get()
-#         outputs['out_file'] = op.abspath(self.inputs.out_file)
-#         return outputs
-
-
-
-def mrtrixcmat(intck, fod_file, roi_volumes, parcellation_scheme, compute_curvature=True, additional_maps={}, output_types=['gPickle'], atlas_info= {}):
-    """ Create the connection matrix using MRTrix fibers and Freesurfer ROIs. """
-    conflow = pe.Workflow(name='MRTRix_connectome_pipeline')
-    #connectome_inputnode = pe.Node(interface=util.IdentityInterface(fields=['intck','fod_file','roi_volumes']),name='inputnode')
-    #connectome_outputnode = pe.Node(interface=util.IdentityInterface(fields=['connectome']),name='outputnode')
-
-    print "INTCK : ",intck
-
-    fibers_filter = pe.Node(interface=FilterTractogram(),name='fibers_filter')
-    fibers_filter.inputs.in_tracks = op.intck
-    fibers_filter.inputs.in_fod = fod_file
-    fibers_filter.inputs.out_file = 'streamlines_weights.txt'
-
-    connectome_builder = pe.Node(interface=BuildConnectome(),name='connectome_builder')
-    connectome_builder.inputs.in_tracks = intck
-    connectome_builder.inputs.in_parc = roi_volumes
-    connectome_builder.inputs.zero_diagonal = True
-
-    conflow.connect([
-                (fibers_filter,connectome_builder,[('out_weights','in_weights')]),
-                ])
-
-    conflow.run()
-
-
 
 def cmat(intrk, roi_volumes, roi_graphmls, parcellation_scheme, compute_curvature=True, additional_maps={}, output_types=['gPickle'], atlas_info = {}):
     """ Create the connection matrix for each resolution using fibers and ROIs. """
@@ -1047,3 +720,644 @@ def cmat(intrk, roi_volumes, roi_graphmls, parcellation_scheme, compute_curvatur
 
     print("Done.")
     print("========================")
+
+
+class rsfmri_conmat_InputSpec(BaseInterfaceInputSpec):
+    func_file = File(exists=True, mandatory=True, desc="fMRI volume")
+    roi_volumes = InputMultiPath(File(exists=True), desc='ROI volumes registered to functional space')
+    roi_graphmls = InputMultiPath(File(exists=True), desc='GraphML description file for ROI volumes (used only if parcellation_scheme == Lausanne2018)')
+    parcellation_scheme = traits.Enum('Lausanne2008',['Lausanne2008','Lausanne2018','NativeFreesurfer','Custom'], usedefault=True)
+    atlas_info = Dict(mandatory = False,desc="custom atlas information")
+    apply_scrubbing = Bool(False)
+    FD = File(exists=True)
+    FD_th = Float()
+    DVARS = File(exists=True)
+    DVARS_th = Float()
+    output_types = traits.List(Str, desc='Output types of the connectivity matrices')
+
+class rsfmri_conmat_OutputSpec(TraitedSpec):
+    avg_timeseries = OutputMultiPath(File(exists=True), desc="ROI average timeseries")
+    scrubbed_idx = File(exists=True)
+    connectivity_matrices = OutputMultiPath(File(exists=True))
+
+class rsfmri_conmat(BaseInterface):
+    input_spec = rsfmri_conmat_InputSpec
+    output_spec = rsfmri_conmat_OutputSpec
+
+    def _run_interface(self,runtime):
+        """ compute the average signal for each GM ROI.
+        """
+        print("Compute average rs-fMRI signal for each cortical ROI")
+        print("====================================================")
+
+        fdata = nib.load( self.inputs.func_file ).get_data()
+
+        tp = fdata.shape[3]
+
+        # OLD
+        # if self.inputs.parcellation_scheme != "Custom":
+        #     resolutions = get_parcellation(self.inputs.parcellation_scheme)
+        # else:
+        #     resolutions = self.inputs.atlas_info
+
+        # NEW
+        print('Parcellation_scheme : %s' % self.inputs.parcellation_scheme)
+
+        if self.inputs.parcellation_scheme != "Custom":
+            if self.inputs.parcellation_scheme != "Lausanne2018":
+                print "get resolutions from parcellation_scheme"
+                resolutions = get_parcellation(self.inputs.parcellation_scheme)
+            else:
+                resolutions = get_parcellation(self.inputs.parcellation_scheme)
+                for parkey, parval in resolutions.items():
+                    for vol, graphml in zip(self.inputs.roi_volumes,self.inputs.roi_graphmls):
+                        print parkey
+                        if parkey in vol:
+                            roi_fname = vol
+                            print roi_fname
+                        if parkey in graphml:
+                            roi_graphml_fname = graphml
+                            print roi_graphml_fname
+                    #roi_fname = roi_volumes[r]
+                    #r += 1
+                    roi       = nib.load(roi_fname)
+                    roiData   = roi.get_data()
+                    resolutions[parkey]['number_of_regions'] = roiData.max()
+                    resolutions[parkey]['node_information_graphml'] = os.path.abspath(roi_graphml_fname)
+
+                del roi, roiData
+                print("##################################################")
+                print("Atlas info (Lausanne2018) :")
+                print(resolutions)
+                print("##################################################")
+        else:
+            print("get resolutions from atlas_info: ")
+            resolutions = self.inputs.atlas_info
+            print(resolutions)
+
+        index = np.linspace(0,tp-1,tp).astype('int')
+
+        # if self.inputs.apply_scrubbing:
+        #     # load scrubbing FD and DVARS series
+        #     FD = np.load( self.inputs.FD )
+        #     DVARS = np.load( self.inputs.DVARS )
+        #     # evaluate scrubbing mask
+        #     FD_th = self.inputs.FD_th
+        #     DVARS_th = self.inputs.DVARS_th
+        #     FD_mask = np.array(np.nonzero(FD < FD_th))[0,:]
+        #     DVARS_mask = np.array(np.nonzero(DVARS < DVARS_th))[0,:]
+        #     index = np.sort(np.unique(np.concatenate((FD_mask,DVARS_mask)))) + 1
+        #     index = np.concatenate(([0],index))
+        #     log_scrubbing = "DISCARDED time points after scrubbing: " + str(FD.shape[0]-index.shape[0]+1) + " over " + str(FD.shape[0]+1)
+        #     print(log_scrubbing)
+        #     np.save( os.path.abspath( 'tp_after_scrubbing.npy'), index )
+        #     sio.savemat( os.path.abspath('tp_after_scrubbing.mat'), {'index':index} )
+        # else:
+        #     index = np.linspace(0,tp-1,tp).astype('int')
+
+        # loop throughout all the resolutions ('scale33', ..., 'scale500')
+        for parkey, parval in resolutions.items():
+            print("Resolution = "+parkey)
+
+            # Open the corresponding ROI
+            print("Open the corresponding ROI")
+            for vol in self.inputs.roi_volumes:
+                if (parkey in vol) or (len(self.inputs.roi_volumes)==1):
+                    roi_fname = vol
+                    print(roi_fname)
+
+            roi = nib.load(roi_fname)
+            mask = roi.get_data()
+
+            ## Compute average time-series
+            # nROIs: number of ROIs for current resolution
+            nROIs = parval['number_of_regions']
+
+            # matrix number of rois vs timepoints
+            ts = np.zeros( (nROIs,tp), dtype = np.float32 )
+
+            # loop throughout all the ROIs (current resolution)
+            for i in range(1,nROIs+1):
+                ts[i-1,:] = fdata[mask==i].mean( axis = 0 )
+            print("ts_shape:",ts.shape)
+
+            np.save( os.path.abspath( 'averageTimeseries_%s.npy' % parkey), ts )
+            sio.savemat( os.path.abspath( 'averageTimeseries_%s.mat' % parkey), {'ts':ts} )
+
+
+        ## Apply scrubbing (if enabled) and compute correlation
+        # loop throughout all the resolutions ('scale33', ..., 'scale500')
+        for parkey, parval in resolutions.items():
+            print("Resolution = "+parkey)
+
+            # Open the corresponding ROI
+            print("Open the corresponding ROI")
+            for vol in self.inputs.roi_volumes:
+                if parkey in vol:
+                    roi_fname = vol
+                    print roi_fname
+            roi       = nib.load(roi_fname)
+            roiData   = roi.get_data()
+
+            #Average roi time-series
+            ts = np.load(os.path.abspath( 'averageTimeseries_%s.npy' % parkey))
+
+            # nROIs: number of ROIs for current resolution
+            nROIs = parval['number_of_regions']
+
+            # Create matrix, add node information from parcellation and recover ROI indexes
+            print("Create the connection matrix (%s rois)" % nROIs)
+            G     = nx.Graph()
+            gp = nx.read_graphml(parval['node_information_graphml'])
+            ROI_idx = []
+            for u,d in gp.nodes(data=True):
+                G.add_node(int(u))
+                for key in d:
+                    G.node[int(u)][key] = d[key]
+                # compute a position for the node based on the mean position of the
+                # ROI in voxel coordinates (segmentation volume )
+                if self.inputs.parcellation_scheme != "Lausanne2018":
+                    G.node[int(u)]['dn_position'] = tuple(np.mean( np.where(mask== int(d["dn_correspondence_id"]) ) , axis = 1))
+                    ROI_idx.append(int(d["dn_correspondence_id"]))
+                else:
+                    G.node[int(u)]['dn_position'] = tuple(np.mean( np.where(mask== int(d["dn_multiscaleID"]) ) , axis = 1))
+                    ROI_idx.append(int(d["dn_multiscaleID"]))
+            # # matrix number of rois vs timepoints
+            # ts = np.zeros( (nROIs,tp), dtype = np.float32 )
+            #
+            # # loop throughout all the ROIs (current resolution)
+            # roi_line = 0
+            # for i in ROI_idx:
+            #     ts[roi_line,:] = fdata[roiData==i].mean( axis = 0 )
+            #     roi_line += 1
+            #
+            # np.save( os.path.abspath('averageTimeseries_%s.npy' % parkey), ts )
+            # sio.savemat( os.path.abspath('averageTimeseries_%s.mat' % parkey), {'TCS':ts} )
+
+            #Censoring time-series
+            if self.inputs.apply_scrubbing:
+                # load scrubbing FD and DVARS series
+                FD = np.load( self.inputs.FD )
+                DVARS = np.load( self.inputs.DVARS )
+                # evaluate scrubbing mask
+                FD_th = self.inputs.FD_th
+                DVARS_th = self.inputs.DVARS_th
+                FD_mask = np.array(np.nonzero(FD < FD_th))[0,:]
+                DVARS_mask = np.array(np.nonzero(DVARS < DVARS_th))[0,:]
+                index = np.sort(np.unique(np.concatenate((FD_mask,DVARS_mask)))) + 1
+                index = np.concatenate(([0],index))
+                log_scrubbing = "DISCARDED time points after scrubbing: " + str(FD.shape[0]-index.shape[0]+1) + " over " + str(FD.shape[0]+1)
+                print(log_scrubbing)
+                np.save( os.path.abspath( 'tp_after_scrubbing.npy'), index )
+                sio.savemat( os.path.abspath('tp_after_scrubbing.mat'), {'index':index} )
+                ts_after_scrubbing = ts[:,index]
+                np.save( os.path.abspath('averageTimeseries_%s_after_scrubbing.npy' % parkey), ts_after_scrubbing )
+                sio.savemat( os.path.abspath('averageTimeseries_%s_after_scrubbing.mat' % parkey), {'ts':ts_after_scrubbing} )
+                ts = ts_after_scrubbing
+                print('ts.shape : ',ts.shape)
+
+                # initialize connectivity matrix
+                nnodes = ts.shape[0]
+                i = -1
+                for i_signal in ts:
+                    i += 1
+                    for j in xrange(i,nnodes):
+                        j_signal = ts[j,:]
+                        value = np.corrcoef(i_signal,j_signal)[0,1]
+                        G.add_edge(ROI_idx[i],ROI_idx[j])
+                        G[ROI_idx[i]][ROI_idx[j]]['corr'] = value
+                        # fmat[i,j] = value
+                        # fmat[j,i] = value
+                # np.save( op.join(gconf.get_timeseries(), 'fconnectome_%s_after_scrubbing.npy' % s), fmat )
+                # sio.savemat( op.join(gconf.get_timeseries(), 'fconnectome_%s_after_scrubbing.mat' % s), {'fmat':fmat} )
+            else:
+                nnodes = ts.shape[0]
+
+                i = -1
+                for i_signal in ts:
+                    i += 1
+                    for j in xrange(i,nnodes):
+                        j_signal = ts[j,:]
+                        value = np.corrcoef(i_signal,j_signal)[0,1]
+                        G.add_edge(ROI_idx[i],ROI_idx[j])
+                        G[ROI_idx[i]][ROI_idx[j]]['corr'] = value
+                        # fmat[i,j] = value
+                        # fmat[j,i] = value
+                # np.save( op.join(gconf.get_timeseries(), 'fconnectome_%s.npy' % s), fmat )
+                # sio.savemat( op.join(gconf.get_timeseries(), 'fconnectome_%s.mat' % s), {'fmat':fmat} )
+
+            # storing network
+            if 'gPickle' in self.inputs.output_types:
+                nx.write_gpickle(G, 'connectome_%s.gpickle' % parkey)
+            if 'mat' in self.inputs.output_types:
+                # edges
+                size_edges = (int(parval['number_of_regions']),int(parval['number_of_regions']))
+                
+                # Get the edge attributes/keys/weights from the first edge and then break. 
+                # Change w.r.t networkx2
+                edge_keys = []
+                for u,v,d in G.edges(data=True):
+                    edge_keys = d.keys()
+                    break
+
+                edge_struct = {}
+                for edge_key in edge_keys:
+                    edge_struct[edge_key] = nx.to_numpy_matrix(G,weight=edge_key)
+
+                # nodes
+                size_nodes = int(parval['number_of_regions'])
+                
+                # Get the node attributes/keys from the first node and then break. 
+                # Change w.r.t networkx2
+                for u,d in G.nodes(data=True):
+                    node_keys = d.keys()
+                    break
+
+                node_struct = {}
+                for node_key in node_keys:
+                    if node_key == 'dn_position':
+                        node_arr = np.zeros([size_nodes,3],dtype=np.float)
+                    else:
+                        node_arr = np.zeros(size_nodes,dtype=np.object_)
+                    node_n = 0
+                    for _,node_data in G.nodes(data=True):
+                        node_arr[node_n] = node_data[node_key]
+                        node_n += 1
+                    node_struct[node_key] = node_arr
+
+                sio.savemat('connectome_%s.mat' % parkey, mdict={'sc':edge_struct,'nodes':node_struct})
+            if 'graphml' in self.inputs.output_types and self.inputs.parcellation_scheme != "Lausanne2018":
+                g2 = nx.Graph()
+                for u_gml,d_gml in G.nodes(data=True):
+                    g2.add_node(u_gml,{'dn_correspondence_id':d_gml['dn_correspondence_id'],
+                                   'dn_fsname':d_gml['dn_fsname'],
+                                   'dn_hemisphere':d_gml['dn_hemisphere'],
+                                   'dn_name':d_gml['dn_name'],
+                                   'dn_position_x':float(d_gml['dn_position'][0]),
+                                   'dn_position_y':float(d_gml['dn_position'][1]),
+                                   'dn_position_z':float(d_gml['dn_position'][2]),
+                                   'dn_region':d_gml['dn_region']})
+                for u_gml,v_gml,d_gml in G.edges(data=True):
+                    g2.add_edge(u_gml,v_gml,{'corr' : float(d_gml['corr'])})
+                nx.write_graphml(g2,'connectome_%s.graphml' % parkey)
+
+            if 'graphml' in self.inputs.output_types and (self.inputs.parcellation_scheme == "Lausanne2018" or self.inputs.parcellation_scheme == 'NativeFreesurfer'):
+                g2 = nx.Graph()
+                for u_gml,d_gml in G.nodes(data=True):
+                    g2.add_node(u_gml,{'dn_multiscaleID':d_gml['dn_multiscaleID'],
+                                   'dn_fsname':d_gml['dn_fsname'],
+                                   'dn_hemisphere':d_gml['dn_hemisphere'],
+                                   'dn_name':d_gml['dn_name'],
+                                   'dn_position_x':float(d_gml['dn_position'][0]),
+                                   'dn_position_y':float(d_gml['dn_position'][1]),
+                                   'dn_position_z':float(d_gml['dn_position'][2]),
+                                   'dn_region':d_gml['dn_region']})
+                for u_gml,v_gml,d_gml in G.edges(data=True):
+                    g2.add_edge(u_gml,v_gml,{'corr' : float(d_gml['corr'])})
+                nx.write_graphml(g2,'connectome_%s.graphml' % parkey)
+
+            if 'cff' in self.inputs.output_types:
+                cvt = cmtk.CFFConverter()
+                cvt.inputs.title = 'Connectome mapper'
+                cvt.inputs.nifti_volumes = self.inputs.roi_volumes
+                cvt.inputs.gpickled_networks = glob.glob(os.path.abspath("connectome_*.gpickle"))
+                cvt.run()
+
+        print("[ DONE ]")
+        return runtime
+
+    def _list_outputs(self):
+        outputs = self._outputs().get()
+        outputs['connectivity_matrices'] = glob.glob(os.path.abspath('connectome*'))
+        outputs['avg_timeseries'] = glob.glob(os.path.abspath('averageTimeseries_*'))
+        if self.inputs.apply_scrubbing:
+            outputs['scrubbed_idx'] = os.path.abspath('tp_after_scrubbing.npy')
+        return outputs
+
+##OLD implementations
+
+# class MRTrix3BaseInputSpec(CommandLineInputSpec):
+#     nthreads = traits.Int(
+#         argstr='-nthreads %d', desc='number of threads. if zero, the number'
+#         ' of available cpus will be used', nohash=True)
+#     # DW gradient table import options
+#     grad_file = File(exists=True, argstr='-grad %s',
+#                      desc='dw gradient scheme (MRTrix format')
+#     grad_fsl = traits.Tuple(
+#         File(exists=True), File(exists=True), argstr='-fslgrad %s %s',
+#         desc='(bvecs, bvals) dw gradient scheme (FSL format')
+#     bval_scale = traits.Enum(
+#         'yes', 'no', argstr='-bvalue_scaling %s',
+#         desc='specifies whether the b - values should be scaled by the square'
+#         ' of the corresponding DW gradient norm, as often required for '
+#         'multishell or DSI DW acquisition schemes. The default action '
+#         'can also be set in the MRtrix config file, under the '
+#         'BValueScaling entry. Valid choices are yes / no, true / '
+#         'false, 0 / 1 (default: true).')
+
+#     in_bvec = File(exists=True, argstr='-fslgrad %s %s',
+#                    desc='bvecs file in FSL format')
+#     in_bval = File(exists=True, desc='bvals file in FSL format')
+
+
+# class MRTrix3Base(CommandLine):
+
+#     def _format_arg(self, name, trait_spec, value):
+#         if name == 'nthreads' and value == 0:
+#             value = 1
+#             try:
+#                 from multiprocessing import cpu_count
+#                 value = cpu_count()
+#             except:
+#                 logger.warn('Number of threads could not be computed')
+#                 pass
+#             return trait_spec.argstr % value
+
+#         if name == 'in_bvec':
+#             return trait_spec.argstr % (value, self.inputs.in_bval)
+
+#         return super(MRTrix3Base, self)._format_arg(name, trait_spec, value)
+
+#     def _parse_inputs(self, skip=None):
+#         if skip is None:
+#             skip = []
+
+#         try:
+#             if (isdefined(self.inputs.grad_file) or
+#                     isdefined(self.inputs.grad_fsl)):
+#                 skip += ['in_bvec', 'in_bval']
+
+#             is_bvec = isdefined(self.inputs.in_bvec)
+#             is_bval = isdefined(self.inputs.in_bval)
+#             if is_bvec or is_bval:
+#                 if not is_bvec or not is_bval:
+#                     raise RuntimeError('If using bvecs and bvals inputs, both'
+#                                        'should be defined')
+#                 skip += ['in_bval']
+#         except AttributeError:
+#             pass
+
+#         return super(MRTrix3Base, self)._parse_inputs(skip=skip)
+
+# class BuildConnectomeInputSpec(CommandLineInputSpec):
+#     in_file = File(exists=True, argstr='%s', mandatory=True, position=-3,
+#                    desc='input tractography')
+#     in_parc = File(exists=True, argstr='%s', position=-2,
+#                    desc='parcellation file')
+#     out_file = File(
+#         'connectome.csv', argstr='%s', mandatory=True, position=-1,
+#         usedefault=True, desc='output file after processing')
+
+#     nthreads = traits.Int(
+#         argstr='-nthreads %d', desc='number of threads. if zero, the number'
+#         ' of available cpus will be used', nohash=True)
+
+#     vox_lookup = traits.Bool(
+#         argstr='-assignment_voxel_lookup',
+#         desc='use a simple voxel lookup value at each streamline endpoint')
+#     search_radius = traits.Float(
+#         argstr='-assignment_radial_search %f',
+#         desc='perform a radial search from each streamline endpoint to locate '
+#         'the nearest node. Argument is the maximum radius in mm; if no node is'
+#         ' found within this radius, the streamline endpoint is not assigned to'
+#         ' any node.')
+#     search_reverse = traits.Float(
+#         argstr='-assignment_reverse_search %f',
+#         desc='traverse from each streamline endpoint inwards along the '
+#         'streamline, in search of the last node traversed by the streamline. '
+#         'Argument is the maximum traversal length in mm (set to 0 to allow '
+#         'search to continue to the streamline midpoint).')
+#     search_forward = traits.Float(
+#         argstr='-assignment_forward_search %f',
+#         desc='project the streamline forwards from the endpoint in search of a'
+#         'parcellation node voxel. Argument is the maximum traversal length in '
+#         'mm.')
+
+#     metric = traits.Enum(
+#         'count', 'meanlength', 'invlength', 'invnodevolume', 'mean_scalar',
+#         'invlength_invnodevolume', argstr='-metric %s', desc='specify the edge'
+#         ' weight metric')
+
+#     in_scalar = File(
+#         exists=True, argstr='-image %s', desc='provide the associated image '
+#         'for the mean_scalar metric')
+
+#     in_weights = File(
+#         exists=True, argstr='-tck_weights_in %s', desc='specify a text scalar '
+#         'file containing the streamline weights')
+
+#     keep_unassigned = traits.Bool(
+#         argstr='-keep_unassigned', desc='By default, the program discards the'
+#         ' information regarding those streamlines that are not successfully '
+#         'assigned to a node pair. Set this option to keep these values (will '
+#         'be the first row/column in the output matrix)')
+#     zero_diagonal = traits.Bool(
+#         argstr='-zero_diagonal', desc='set all diagonal entries in the matrix '
+#         'to zero (these represent streamlines that connect to the same node at'
+#         ' both ends)')
+
+
+# class BuildConnectomeOutputSpec(TraitedSpec):
+#     out_file = File(exists=True, desc='the output response file')
+
+
+# class BuildConnectome(MRTrix3Base):
+
+#     """
+#     Generate a connectome matrix from a streamlines file and a node
+#     parcellation image
+#     Example
+#     -------
+#     >>> import nipype.interfaces.mrtrix3 as mrt
+#     >>> mat = mrt.BuildConnectome()
+#     >>> mat.inputs.in_file = 'tracks.tck'
+#     >>> mat.inputs.in_parc = 'aparc+aseg.nii'
+#     >>> mat.cmdline                               # doctest: +ELLIPSIS +ALLOW_UNICODE
+#     'tck2connectome tracks.tck aparc+aseg.nii connectome.csv'
+#     >>> mat.run()                                 # doctest: +SKIP
+#     """
+
+#     _cmd = 'tck2connectome'
+#     input_spec = BuildConnectomeInputSpec
+#     output_spec = BuildConnectomeOutputSpec
+
+#     def _list_outputs(self):
+#         outputs = self.output_spec().get()
+#         outputs['out_file'] = op.abspath(self.inputs.out_file)
+#         return outputs
+
+def mrtrixcmat(intck, fod_file, roi_volumes, parcellation_scheme, compute_curvature=True, additional_maps={}, output_types=['gPickle'], atlas_info= {}):
+    """ Create the connection matrix using MRTrix fibers and Freesurfer ROIs. """
+    conflow = pe.Workflow(name='MRTRix_connectome_pipeline')
+    #connectome_inputnode = pe.Node(interface=util.IdentityInterface(fields=['intck','fod_file','roi_volumes']),name='inputnode')
+    #connectome_outputnode = pe.Node(interface=util.IdentityInterface(fields=['connectome']),name='outputnode')
+
+    print "INTCK : ",intck
+
+    fibers_filter = pe.Node(interface=FilterTractogram(),name='fibers_filter')
+    fibers_filter.inputs.in_tracks = op.intck
+    fibers_filter.inputs.in_fod = fod_file
+    fibers_filter.inputs.out_file = 'streamlines_weights.txt'
+
+    connectome_builder = pe.Node(interface=BuildConnectome(),name='connectome_builder')
+    connectome_builder.inputs.in_tracks = intck
+    connectome_builder.inputs.in_parc = roi_volumes
+    connectome_builder.inputs.zero_diagonal = True
+
+    conflow.connect([
+                (fibers_filter,connectome_builder,[('out_weights','in_weights')]),
+                ])
+
+    conflow.run()
+
+
+#CMP2
+def prob_cmat(intrk, roi_volumes, parcellation_scheme, output_types=['gPickle'], atlas_info = {}):
+
+    print("Filling probabilistic connectivity matrices:")
+
+    if parcellation_scheme != "Custom":
+        resolutions = get_parcellation(parcellation_scheme)
+    else:
+        resolutions = atlas_info
+
+    firstROIFile = roi_volumes[0]
+    firstROI = nibabel.load(firstROIFile)
+    roiVoxelSize = firstROI.get_header().get_zooms()
+
+    for parkey, parval in resolutions.items():
+        print("Resolution = "+parkey)
+
+        # Open the corresponding ROI
+        print("Open the corresponding ROI")
+        for vol in roi_volumes:
+            if parkey in vol:
+                roi_fname = vol
+                print roi_fname
+        roi       = nibabel.load(roi_fname)
+        roiData   = roi.get_data()
+
+        # Create the matrix
+        nROIs = parval['number_of_regions']
+        print("Create the connection matrix (%s rois)" % nROIs)
+        G     = nx.Graph()
+
+        # add node information from parcellation
+        gp = nx.read_graphml(parval['node_information_graphml'])
+        for u,d in gp.nodes(data=True):
+            G.add_node(int(u))
+            for key in d:
+                G.node[int(u)][key] = d[key]
+            # compute a position for the node based on the mean position of the
+            # ROI in voxel coordinates (segmentation volume )
+            G.node[int(u)]['dn_position'] = tuple(np.mean( np.where(roiData== int(d["dn_correspondence_id"]) ) , axis = 1))
+
+        graph_matrix = np.zeros((nROIs,nROIs),dtype = int)
+
+        pc = -1
+
+        for intrk_i in range(0,len(intrk)):
+            pcN = int(round( float(100*intrk_i)/len(intrk) ))
+            if pcN > pc and pcN%20 == 0:
+                pc = pcN
+                print('%4.0f%%' % (pc))
+            fib, hdr    = nibabel.trackvis.read(intrk[intrk_i], False)
+            (endpoints,endpointsmm) = create_endpoints_array(fib, roiVoxelSize, False)
+            n = len(fib)
+
+            dis = 0
+
+            pc = -1
+            for i in range(n):  # n: number of fibers
+
+                # ROI start => ROI end
+                try:
+                    startROI = int(roiData[endpoints[i, 0, 0], endpoints[i, 0, 1], endpoints[i, 0, 2]]) # endpoints from create_endpoints_array
+                    endROI   = int(roiData[endpoints[i, 1, 0], endpoints[i, 1, 1], endpoints[i, 1, 2]])
+                except IndexError:
+                    print("An index error occured for fiber %s. This means that the fiber start or endpoint is outside the volume. Continue." % i)
+                    continue
+
+                # Filter
+                if startROI == 0 or endROI == 0:
+                    dis += 1
+                    #fiberlabels[i,0] = -1
+                    continue
+
+                if startROI > nROIs or endROI > nROIs:
+        #                print("Start or endpoint of fiber terminate in a voxel which is labeled higher")
+        #                print("than is expected by the parcellation node information.")
+        #                print("Start ROI: %i, End ROI: %i" % (startROI, endROI))
+        #                print("This needs bugfixing!")
+                    continue
+
+                # Update fiber label
+                # switch the rois in order to enforce startROI < endROI
+                if endROI < startROI:
+                    tmp = startROI
+                    startROI = endROI
+                    endROI = tmp
+
+                # Add edge to graph
+                if G.has_edge(startROI, endROI):
+                    G[startROI][endROI]['n_tracks'] += 1
+                else:
+                    G.add_edge(startROI, endROI, n_tracks  = 1)
+
+                graph_matrix[startROI-1][endROI-1] +=1
+
+        tot_tracks = graph_matrix.sum()
+
+        for u,v,d in G.edges(data=True):
+            G.remove_edge(u,v)
+            di = { 'number_of_fibers' : (float(d['n_tracks']) / tot_tracks.astype(float))}
+            G.add_edge(u,v)
+            for key in di:
+                G[u][v][key] = di[key]
+
+        # storing network
+        if 'gPickle' in output_types:
+            nx.write_gpickle(G, 'connectome_%s.gpickle' % parkey)
+        if 'mat' in output_types:
+            # edges
+            size_edges = (parval['number_of_regions'],parval['number_of_regions'])
+            edge_keys = G.edges(data=True)[0][2].keys()
+
+            edge_struct = {}
+            for edge_key in edge_keys:
+                edge_struct[edge_key] = nx.to_numpy_matrix(G,weight=edge_key)
+
+            # nodes
+            size_nodes = parval['number_of_regions']
+            node_keys = G.nodes(data=True)[0][1].keys()
+
+            node_struct = {}
+            for node_key in node_keys:
+                if node_key == 'dn_position':
+                    node_arr = np.zeros([size_nodes,3],dtype=np.float)
+                else:
+                    node_arr = np.zeros(size_nodes,dtype=np.object_)
+                node_n = 0
+                for _,node_data in G.nodes(data=True):
+                    node_arr[node_n] = node_data[node_key]
+                    node_n += 1
+                node_struct[node_key] = node_arr
+
+            scipy.io.savemat('connectome_%s.mat' % parkey, mdict={'sc':edge_struct,'nodes':node_struct})
+        if 'graphml' in output_types:
+            g2 = nx.Graph()
+            for u_gml,v_gml,d_gml in G.edges(data=True):
+                g2.add_edge(u_gml,v_gml)
+                for key in d_gml:
+                    g2[u_gml][v_gml][key] = d_gml[key]
+            for u_gml,d_gml in G.nodes(data=True):
+                g2.add_node(u_gml)
+                g2.node[u_gml]['dn_correspondence_id'] = d_gml['dn_correspondence_id']
+                g2.node[u_gml]['dn_fsname'] = d_gml['dn_fsname']
+                g2.node[u_gml]['dn_hemisphere'] = d_gml['dn_hemisphere']
+                g2.node[u_gml]['dn_name'] = d_gml['dn_name']
+                g2.node[u_gml]['dn_position_x'] = d_gml['dn_position'][0]
+                g2.node[u_gml]['dn_position_y'] = d_gml['dn_position'][1]
+                g2.node[u_gml]['dn_position_z'] = d_gml['dn_position'][2]
+                g2.node[u_gml]['dn_region'] = d_gml['dn_region']
+            nx.write_graphml(g2,'connectome_%s.graphml' % parkey)
